Add files via upload
Lab 4
The purpose of this lab was to determine suitable sites for wind energy in the Philippines. 
The data includes five raster layers: urban areas, water bodies, protected areas, slope, and mean annual wind speed. 
A moving window was created to calculate the mean values of pixels iteratively. 
Each layer was turned into a Boolean array based on the suitable criteria assessed for each variable. 
All 5 Boolean arrays were combined, and only sites that resulted in a value of 5 were considered suitable.
This array was then converted into a GeoTiff. 
Distance was then calculated between the suitable sites and the closest transmission stations.