Add files via upload
Lab 2

Three text files were given to be encoded into the boundaries of three districts in eastern Tanzania. 
These were then coded to have two more variables, the number of vertices (or coordinate pairs) and the district each vertices was a part of. 
A Boolean array was then used to calculate the agricultural land within each district. This was then turned into a percentage of the total land in each district.