Add files via upload
Final Project

Final project for the class. 

Takes temperature data gathered from a Utility Scale Solar Array in Boulder, Colorado, over the span of about 4 months. Includes 18 sensors, with data taken hourly. 

The data existed in two sets of .csv files, resulting in 36 files for 18 sensors. These files were merged based on unique sensor identifiers, and then null values were filled in. 
These null values were then summed(result was that none existed).
Finally, the formatting of the data frames was altered to group them by day, and to then find the minimum and maximum temperature values based on day, without losing the time stamp of that measurement. The data was converted to a numpy array for easier visualization, and then output as a .csv file.